FormWorkoutSession works as expectec.

+++ FittSoft/FittSoft/WorkoutSession.cs
@@ -10,6 +10,6 @@ namespace FittSoft
     {
         public DateTime startTime { get; set; }
         public DateTime endTime { get; set; }
-        public int duration { get; set; }
+        public decimal duration { get; set; }
     }
 }
